fix tests to meet requirements
